MoveSideListItemWrapper to a new Component
