+++ src/SideList/SideList.js
@@ -3,9 +3,6 @@ import { useSpring, animated } from 'react-spring'
 import PropTypes from 'prop-types'
 import styled from 'styled-components/macro'
 
-import SongListItem from './SongListItem'
-import SetlistItem from './SetlistItem'
-import AddSetlist from '../forms/AddSetlist'
 import ListMenu from './ListMenu'
 import SideListTitle from './SideListTitle'
 import InputField from '../forms/InputField'
@@ -17,6 +14,7 @@ import useSongs from '../hooks/useSongs'
 import { patchSetlist } from '../services.js'
 
 import search from '../icons/search-blue.svg'
+import SideListItemWrapper from './SideListItemWrapper'
 
 function SideList({
   songs,
@@ -82,9 +80,6 @@ function SideList({
     }
   }, [searchInput, sideListType, songs, setlists])
 
-  let sideListContent
-  handleSideListType()
-
   const AnimatedSideListWrapperBorder = animated(Container)
   const flyIn = useSpring({
     width: isSideListShown ? dimensions.sideListWidth + 'px' : '0px',
@@ -115,7 +110,29 @@ function SideList({
               style={{ margin: dimensions.sideListPadding + 'px' }}
             ></InputField>
           )}
-          <SideListItemWrapper>{sideListContent}</SideListItemWrapper>
+          <SideListItemWrapper
+            songs={songs}
+            setlists={setlists}
+            setSetlists={setSetlists}
+            setlistSongs={setlistSongs}
+            setSetlistSongs={setSetlistSongs}
+            setSearchInput={setSearchInput}
+            sideListType={sideListType}
+            swipeIndex={swipeIndex}
+            setSwipeIndex={setSwipeIndex}
+            showSearchField={showSearchField}
+            fuzzySearchResult={fuzzySearchResult}
+            handleChangeIndex={handleChangeIndex}
+            setlistsIsLoading={setlistsIsLoading}
+            setSideListType={setSideListType}
+            setActiveSetlist={setActiveSetlist}
+            activeSetlist={activeSetlist}
+            setSetlistsIsLoading={setSetlistsIsLoading}
+            sideListTitle={sideListTitle}
+            setSideListTitle={setSideListTitle}
+            activeSetlistIndex={activeSetlistIndex}
+          ></SideListItemWrapper>
+          {/* <SideListItemWrapper>{sideListContent}</SideListItemWrapper> */}
         </ItemSearchWrapper>
         <ListMenu
           sideListType={sideListType}
@@ -125,66 +142,12 @@ function SideList({
           setActiveSetlist={setActiveSetlist}
           setSetlists={setSetlists}
           setlists={setlists}
-          setlistSongs={setlistSongs}
           handleSaveSongsToSetlist={handleSaveSongsToSetlist}
           setKeyCounter={setKeyCounter}
         />
       </SideListWrapperBorder>
     </AnimatedSideListWrapperBorder>
   )
-
-  function handleIsSongsShown(songs) {
-    const songList = showSearchField ? fuzzySearchResult : songs
-    if (songs) {
-      sideListContent = songList.map(song => (
-        <SongListItem
-          key={song._id}
-          index={findSongIndex(song._id)}
-          sideListType={sideListType}
-          song={song}
-          swipeIndex={swipeIndex}
-          handleChangeIndex={handleChangeIndex}
-          setlistSongs={setlistSongs}
-          setSetlistSongs={setSetlistSongs}
-          setSearchInput={setSearchInput}
-        />
-      ))
-    } else {
-      sideListContent = 'no song'
-    }
-  }
-
-  function handleIsSetListsShown(setlists) {
-    const setlistList = showSearchField ? fuzzySearchResult : setlists
-    if (setlists) {
-      sideListContent = setlistsIsLoading
-        ? 'loading...'
-        : setlistList.map(setlist => (
-            <SetlistItem
-              key={setlist._id}
-              setlist={setlist}
-              sideListType={sideListType}
-              setSideListType={setSideListType}
-              setActiveSetlist={setActiveSetlist}
-              activeSetlist={activeSetlist}
-              setSwipeIndex={setSwipeIndex}
-              setSetlistSongs={setSetlistSongs}
-            />
-          ))
-    } else {
-      sideListContent = 'no setlists'
-    }
-  }
-
-  function handleIsAddSetlistShown() {
-    sideListContent = (
-      <AddSetlist
-        setSideListType={setSideListType}
-        setSetlists={setSetlists}
-        setSetlistsIsLoading={setSetlistsIsLoading}
-      />
-    )
-  }
 
   async function handleSaveSongsToSetlist() {
     const index = setlists.findIndex(setlist => setlist._id === activeSetlist)
@@ -221,48 +184,7 @@ function SideList({
     }
     return tokens.join('')
   }
-
-  function findSongIndex(songID) {
-    const songList =
-      sideListType === 'singleSetlist'
-        ? setlists[activeSetlistIndex].songs
-        : songs
-    return songList.findIndex(song => song._id === songID)
-  }
-
-  function handleSideListType() {
-    if (sideListType === 'allSongs') {
-      handleIsSongsShown(songs)
-      sideListTitle === 'All Songs' || setSideListTitle('All Songs')
-    } else if (sideListType === 'setlists') {
-      handleIsSetListsShown(setlists)
-      sideListTitle === 'All Sets' || setSideListTitle('All Sets')
-    } else if (sideListType === 'singleSetlist') {
-      handleIsSongsShown(setlists[activeSetlistIndex].songs)
-      sideListTitle === setlists[activeSetlistIndex].setlistName ||
-        setSideListTitle(setlists[activeSetlistIndex].setlistName)
-    } else if (sideListType === 'addSetlist') {
-      handleIsAddSetlistShown()
-      sideListTitle === 'Add Set' || setSideListTitle('Add Set')
-    } else if (sideListType === 'addSongToSetlist') {
-      handleIsSongsShown(songs)
-      sideListTitle === 'Add Song to Set' || setSideListTitle('Add Song to Set')
-    }
-  }
 }
-
-const SideListItemWrapper = styled.ul`
-  justify-self: stretch;
-  align-self: center;
-  list-style: none;
-  padding: ${dimensions.sideListPadding + 'px'};
-  height: 100%;
-  overflow: auto;
-  -webkit-user-select: none;
-  -moz-user-select: none;
-  -ms-user-select: none;
-  user-select: none;
-`
 
 const ItemSearchWrapper = styled.div`
   overflow: scroll;
